python知识与作业

+++ 14、练习的项目/11_类Fabric主机管理程序/fabric_hosts/core/db_handler.py
@@ -34,6 +34,8 @@ def delete_host(host_id):
 def query_host(host_id):
     """查询单个主机信息"""
     json_file = os.path.join(DB_Host, "{0}.json".format(host_id))
+    if not os.path.exists(json_file):
+        raise SomeError("id编号为{0}的主机不存在".format(host_id))
     with open(json_file, "r") as f:
         return json.loads(f.read())
 
@@ -48,6 +50,8 @@ def list_hosts():
 def save_host_group(host_group):
     """添加/修改后，保存主机组信息"""
     json_file = os.path.join(DB_HostGroups, "{0}.json".format(host_group.group_id))
+    if os.path.exists(json_file):
+        raise SomeError("id编号为{0}的主机组已存在".format(host_group.group_id))
     with open(json_file, "w") as f:
         f.write(json.dumps(host_group.__dict__))
 
@@ -72,6 +76,8 @@ def delete_host_group(group_id):
 def query_host_group(group_id):
     """查询单个主机组信息"""
     json_file = os.path.join(DB_HostGroups, "{0}.json".format(group_id))
+    if not os.path.exists(json_file):
+        raise SomeError("id编号为{0}的主机组不存在".format(group_id))
     with open(json_file, "r") as f:
         return json.loads(f.read())
 
@@ -84,12 +90,10 @@ def list_host_groups():
 
 
 def get_new_host_id():
-    """新增时，生成新主机id编号"""
-    num = len(list_hosts())
-    return "hostCode" + str(num)
+    """新增时，生成新主机id编号，从1000开始计数"""
+    return str(1000 + len(list_hosts()))
 
 
 def get_new_host_group_id():
-    """新增时，生成新主机组id编号"""
-    num = len(list_host_groups())
-    return "hostGroupCode" + str(num)
+    """新增时，生成新主机组id编号，从1000开始计数"""
+    return str(1000 + len(list_host_groups()))
